reports_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
clickhouse-driver==0.2.9
PyJWT==2.9.0
